requirements.txt
langchain~=0.3.25
openai~=0.27.8
pathlib2~=2.3.7
python-dotenv~=1.1.0